+++ src/integrations/lottie/render.rs
@@ -16,7 +16,7 @@ use super::{
 use crate::{
     SkipEncoding, VelloScreenSpace,
     render::{
-        VelloEntityCountData, VelloView,
+        NoVelloScale, VelloEntityCountData, VelloScreenScale, VelloView, VelloWorldScale,
         prepare::{PrepareRenderInstance, PreparedAffine, PreparedTransform},
     },
 };
@@ -31,6 +31,7 @@ pub struct ExtractedLottieAsset {
     pub theme: Option<Theme>,
     pub playhead: f64,
     pub screen_space: Option<VelloScreenSpace>,
+    pub no_scaling: Option<NoVelloScale>,
 }
 
 pub fn extract_lottie_assets(
@@ -52,6 +53,7 @@ pub fn extract_lottie_assets(
                 &ViewVisibility,
                 &InheritedVisibility,
                 Option<&VelloScreenSpace>,
+                Option<&NoVelloScale>,
             ),
             Without<SkipEncoding>,
         >,
@@ -76,6 +78,7 @@ pub fn extract_lottie_assets(
         view_visibility,
         inherited_visibility,
         screen_space,
+        no_scaling,
     ) in query_vectors.iter()
     {
         // Skip if visibility conditions are not met
@@ -102,6 +105,7 @@ pub fn extract_lottie_assets(
                     alpha: asset.alpha,
                     ui_node: ui_node.cloned(),
                     screen_space: screen_space.cloned(),
+                    no_scaling: no_scaling.cloned(),
                 })
                 .insert(TemporaryRenderEntity);
             n_lotties += 1;
@@ -115,13 +119,21 @@ pub fn prepare_asset_affines(
     mut commands: Commands,
     views: Query<(&ExtractedCamera, &ExtractedView), With<Camera2d>>,
     mut render_entities: Query<(Entity, &ExtractedLottieAsset)>,
+    world_scale: Res<VelloWorldScale>,
+    screen_scale: Res<VelloScreenScale>,
 ) {
     for (camera, view) in views.iter() {
         let viewport_size: UVec2 = camera.physical_viewport_size.unwrap();
         for (entity, render_entity) in render_entities.iter_mut() {
             // Prepare render data needed for the subsequent render system
             let final_transform = render_entity.final_transform();
-            let affine = render_entity.scene_affine(view, *final_transform, viewport_size);
+            let affine = render_entity.scene_affine(
+                view,
+                *final_transform,
+                viewport_size,
+                world_scale.0,
+                screen_scale.0,
+            );
 
             commands.entity(entity).insert((affine, final_transform));
         }
@@ -142,11 +154,18 @@ impl PrepareRenderInstance for ExtractedLottieAsset {
         view: &ExtractedView,
         world_transform: GlobalTransform,
         viewport_size: UVec2,
+        world_scale: f32,
+        screen_scale: f32,
     ) -> PreparedAffine {
         let local_center_matrix = self.asset.local_transform_center.compute_matrix().inverse();
 
         let raw_transform = if let Some(node) = self.ui_node {
             let mut model_matrix = world_transform.compute_matrix();
+
+            if self.no_scaling.is_none() {
+                model_matrix.x_axis.x *= screen_scale;
+                model_matrix.y_axis.y *= screen_scale;
+            }
 
             let asset_size = Vec2::new(self.asset.width, self.asset.height);
 
@@ -161,7 +180,13 @@ impl PrepareRenderInstance for ExtractedLottieAsset {
             local_center_matrix.w_axis.y *= -1.0;
             model_matrix * local_center_matrix
         } else if self.screen_space.is_some() {
-            let model_matrix = world_transform.compute_matrix();
+            let mut model_matrix = world_transform.compute_matrix();
+
+            if self.no_scaling.is_none() {
+                model_matrix.x_axis.x *= screen_scale;
+                model_matrix.y_axis.y *= screen_scale;
+            }
+
             let mut local_center_matrix = local_center_matrix;
             local_center_matrix.w_axis.y *= -1.0;
             model_matrix * local_center_matrix
@@ -179,6 +204,11 @@ impl PrepareRenderInstance for ExtractedLottieAsset {
 
             let mut model_matrix = world_transform.compute_matrix() * local_matrix;
             model_matrix.w_axis.y *= -1.0;
+
+            if self.no_scaling.is_none() {
+                model_matrix.x_axis.x *= world_scale;
+                model_matrix.y_axis.y *= world_scale;
+            }
 
             let (projection_mat, view_mat) = {
                 let mut view_mat = view.world_from_view.compute_matrix();